requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.28
botocore==1.38.28
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.10
fastapi==0.115.12
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8
langsmith==0.3.44
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
narwhals==1.41.0
numpy==2.2.6
oauthlib==3.2.2
openai==1.83.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
phidata==2.7.10
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.11.12
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
websocket-client==1.8.0
yarl==1.20.0
zstandard==0.23.0
